+++ Microsoft.Azure.Cosmos/src/CosmosClient.cs
@@ -722,18 +722,12 @@ namespace Microsoft.Azure.Cosmos
 
         private HttpClientHandler CreateHttpClientHandler(CosmosClientOptions clientOptions)
         {
-            if (clientOptions == null || (!clientOptions.DisableSSLCertificateVerification && clientOptions.WebProxy == null))
+            if (clientOptions == null || (clientOptions.WebProxy == null))
             {
                 return null;
             }
 
             HttpClientHandler httpClientHandler = new HttpClientHandler();
-
-            if (clientOptions.DisableSSLCertificateVerification)
-            {
-                httpClientHandler.ClientCertificateOptions = ClientCertificateOption.Manual;
-                httpClientHandler.ServerCertificateCustomValidationCallback += (sender, cert, chain, sslPolicyErrors) => true;
-            }
 
             if (clientOptions.WebProxy != null)
             {